donw

+++ IBF.py
@@ -1,9 +1,6 @@
 ###  Import Module
-import os
-import time
-import urllib.request
+import os, os.path, time, urllib.request
 from PIL import Image
-from io import BytesIO
 from selenium import webdriver
 from selenium.webdriver.chrome.options import Options
 from selenium.webdriver.common.by import By
@@ -20,11 +17,14 @@ except FileExistsError:
     pass
 folderIBF = ("D:/TAHUN "+thnskr+"/"+str (bulanangka)+". "+bulan+"/IBF/")
 src = "https://web.meteo.bmkg.go.id//media/data/bmkg/ibfnew/34_papua_24.png"
-urllib.request.urlretrieve(src, folderIBF+bsk+".png")
-image = os.path.realpath(folderIBF+bsk+".png")
+if os.path.exists(folderIBF+bsk+".png"):
+    pass
+else:
+    urllib.request.urlretrieve(src, folderIBF+bsk+".png")
+image = folderIBF+bsk+".png"
 narasi = ("Informasi Prakiraan Cuaca Berbasis Dampak wilayah Provinsi Papua berlaku "+tanggalbesok+" pukul 09.00 WIT - "+tanggallusa+" pukul 09.00 WIT Informasi lebih lengkap dapat diakses pada : http://signature.bmkg.go.id/")
 
-## Setting Chrome Driver
+### Setting Driver
 import pathlib
 cPath = pathlib.Path(__file__).parent.resolve()
 currentPath = str(cPath)
@@ -34,34 +34,22 @@ chromeOptions.add_argument("user-data-dir=" + currentPath + "cookies")
 #chromeOptions.add_argument("--headless")
 driver = webdriver.Chrome(driverPath, options=chromeOptions)
 
+## Whatsapp
 driver.get("http://web.whatsapp.com")
-WebDriverWait(driver, 50).until(EC.presence_of_element_located((By.XPATH, '//*[@id="side"]/div[1]/div/div/div[2]/div/div[2]'))).send_keys("Arif Ripcy", Keys.ENTER)
-WebDriverWait(driver, 50).until(EC.presence_of_element_located((By.XPATH, '//*[@id="main"]/footer//*[@data-icon="clip"]/..'))).click()
-WebDriverWait(driver, 50).until(EC.presence_of_element_located((By.XPATH, '//*[@id="main"]/footer//*[@data-icon="attach-image"]/../input'))).send_keys(image)
-WebDriverWait(driver, 50).until(EC.presence_of_element_located((By.XPATH, '//*[@id="app"]/div/div/div[2]/div[2]/span/div/span/div/div/div[2]/div/div[1]/div[3]/div/div/div[2]/div[1]/div[2]'))).send_keys(narasi)
-WebDriverWait(driver, 50).until(EC.presence_of_element_located((By.XPATH, '//*[@id="app"]/div/div/div[2]/div[2]/span/div/span/div/div/div[2]/div/div[2]/div[2]/div'))).click()
+WebDriverWait(driver, 200).until(EC.presence_of_element_located((By.XPATH, '//*[@id="side"]/div[1]/div/div/div[2]/div/div[2]'))).send_keys("Arif", Keys.ENTER)
+WebDriverWait(driver, 200).until(EC.presence_of_element_located((By.XPATH, '//*[@id="main"]/footer//*[@data-icon="clip"]/..'))).click()
+WebDriverWait(driver, 200).until(EC.presence_of_element_located((By.XPATH, '//*[@id="main"]/footer//*[@data-icon="attach-image"]/../input'))).send_keys(image)
+WebDriverWait(driver, 200).until(EC.presence_of_element_located((By.XPATH, '//*[@id="app"]/div/div/div[2]/div[2]/span/div/span/div/div/div[2]/div/div[1]/div[3]/div/div/div[2]/div[1]/div[2]'))).send_keys(narasi)
+WebDriverWait(driver, 200).until(EC.presence_of_element_located((By.XPATH, '//*[@id="app"]/div/div/div[2]/div[2]/span/div/span/div/div/div[2]/div/div[2]/div[2]/div'))).click()
 time.sleep(10)
-'''
-import requests
 
-url = "https://api.telegram.org/bot5663329448:AAFYtirH9HbrgcsJWPisd-blu-AL1jV0Aes/sendPhoto"
-
-payload = {
-    "photo": image,
-    "caption": narasi,
-    "disable_notification": False,
-    "reply_to_message_id": None
-}
-headers = {
-    "accept": "application/json",
-    "content-type": "application/json"
-}
-
-response = requests.post(url, json=payload, headers=headers)
-
-print(response.text)
-
-'''
+##Telegram
+import telegram
+from telegram import InputMediaPhoto as imp
+apiToken = '5663329448:AAFYtirH9HbrgcsJWPisd-blu-AL1jV0Aes'
+chatID = '40071216'
+bot = telegram.Bot(token=apiToken)
+bot.send_media_group(chat_id=chatID,media=[imp(open(image, 'rb'), caption=narasi)])
 
 
 driver.quit()
